refactor(refactor-patrol): extract ThesisNormalizer from Reviewer
The Reviewer mixed four roles: per-feature orchestration, raw-thesis
normalization (alias/evidence repair), admissibility+guidance policy, and
agent plumbing. Dogfooding showed the normalization/policy block is the
fastest-evolving part, so it moves to a pure ThesisNormalizer — no I/O, no
agent coupling — returning a Thesis or an Invalid carrying schema errors;
the Reviewer keeps orchestration and error recording. Behavior unchanged.

The two thesis-consuming test files now share one canonical fixture module
(thesis_fixtures.rb) instead of ~60 duplicated lines, and normalization
policy is unit-tested directly in thesis_normalizer_test.rb.

Live-tested: real-agent dry run against the dogfood project returned both
theses admissible with evidence intact (autoreview pass: 8 finder angles,
no confirmed correctness findings; fixture duplication fixed).
